refactor(dto): rbf, cpfp 관련 dto 클래스 중복 제거
- 디렉터리 변경
- 네이밍 변경
